+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## [Unreleased]
+
+## [0.6.0] - 2022-06-17
 ### Added
 - Copy method for `Translator.copy(**overrides)`.
 
@@ -82,7 +84,8 @@ First release on PyPI!
 - The `rics.mapping` package for linking elements in multiple directions
 - The `rics.cardinality` package; enum types for `1:1`, `1:N`, `N:1`, and `M:N`
 
-[Unreleased]: https://github.com/rsundqvist/rics/compare/v0.5.0...HEAD
+[Unreleased]: https://github.com/rsundqvist/rics/compare/v0.6.0...HEAD
+[0.6.0]: https://github.com/rsundqvist/rics/compare/v0.5.0...v0.6.0
 [0.5.0]: https://github.com/rsundqvist/rics/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/rsundqvist/rics/compare/v0.3.2...v0.4.0
 [0.3.2]: https://github.com/rsundqvist/rics/compare/v0.3.1...v0.3.2